DeleteBookmark() will check the response from pinboard and return errors.
The 'delete' command has been adjusted to check if any errors occurred when
deleting, it will exit with a failure status then. Also you can supply the URLs
to delete directly as arguments to `delete`.

Internally the code has been streamlined to use the functions alredy existing in the
pinboard package.